+++ main/sketch.cpp
@@ -25,7 +25,7 @@ limitations under the License.
 // README FIRST, README FIRST, README FIRST
 //
 // Bluepad32 has a built-in interactive console.
-// By default it is enabled (hey, this is a great feature!).
+// By default, it is enabled (hey, this is a great feature!).
 // But it is incompatible with Arduino "Serial" class.
 //
 // Instead of using "Serial" you can use Bluepad32 "Console" class instead.
@@ -81,7 +81,7 @@ void dumpGamepad(ControllerPtr ctl) {
         "idx=%d, dpad: 0x%02x, buttons: 0x%04x, axis L: %4d, %4d, axis R: %4d, %4d, brake: %4d, throttle: %4d, "
         "misc: 0x%02x, gyro x:%6d y:%6d z:%6d, accel x:%6d y:%6d z:%6d\n",
         ctl->index(),        // Controller Index
-        ctl->dpad(),         // DPAD
+        ctl->dpad(),         // D-pad
         ctl->buttons(),      // bitmask of pressed buttons
         ctl->axisX(),        // (-511 - 512) left X Axis
         ctl->axisY(),        // (-511 - 512) left Y axis
@@ -89,7 +89,7 @@ void dumpGamepad(ControllerPtr ctl) {
         ctl->axisRY(),       // (-511 - 512) right Y axis
         ctl->brake(),        // (0 - 1023): brake button
         ctl->throttle(),     // (0 - 1023): throttle (AKA gas) button
-        ctl->miscButtons(),  // bitmak of pressed "misc" buttons
+        ctl->miscButtons(),  // bitmask of pressed "misc" buttons
         ctl->gyroX(),        // Gyro X
         ctl->gyroY(),        // Gyro Y
         ctl->gyroZ(),        // Gyro Z
@@ -173,6 +173,8 @@ void processGamepad(ControllerPtr ctl) {
     // Another way to query controller data is by getting the buttons() function.
     // See how the different "dump*" functions dump the Controller info.
     dumpGamepad(ctl);
+
+    // See ArduinoController.h for all the available functions.
 }
 
 void processMouse(ControllerPtr ctl) {
@@ -222,6 +224,25 @@ void processBalanceBoard(ControllerPtr ctl) {
     dumpBalanceBoard(ctl);
 }
 
+void processControllers() {
+    for (auto myController : myControllers) {
+        if (myController && myController->isConnected() && myController->hasData()) {
+            if (myController->isGamepad()) {
+                processGamepad(myController);
+            } else if (myController->isMouse()) {
+                processMouse(myController);
+            } else if (myController->isKeyboard()) {
+                processKeyboard(myController);
+            } else if (myController->isBalanceBoard()) {
+                processBalanceBoard(myController);
+            } else {
+                Console.printf("Data not available yet\n");
+                continue;
+            }
+        }
+    }
+}
+
 // Arduino setup function. Runs in CPU 1
 void setup() {
     Console.printf("Firmware: %s\n", BP32.firmwareVersion());
@@ -235,52 +256,31 @@ void setup() {
     // a "device factory reset", or similar.
     // Calling "forgetBluetoothKeys" in setup() just as an example.
     // Forgetting Bluetooth keys prevents "paired" gamepads to reconnect.
-    // But might also fix some connection / re-connection issues.
+    // But it might also fix some connection / re-connection issues.
     BP32.forgetBluetoothKeys();
 
     // Enables mouse / touchpad support for gamepads that support them.
-    // When enabled controllers like DualSense and DualShock4 generate two connected devices:
+    // When enabled, controllers like DualSense and DualShock4 generate two connected devices:
     // - First one: the gamepad
-    // - Second one, which is a "vritual device", is a mouse
-    // By default it is disabled.
+    // - Second one, which is a "virtual device", is a mouse.
+    // By default, it is disabled.
     BP32.enableVirtualDevice(false);
 }
 
-// Arduino loop function. Runs in CPU 1
+// Arduino loop function. Runs in CPU 1.
 void loop() {
-    // This call fetches all the gamepad info from the NINA (ESP32) module.
-    // Just call this function in your main loop.
-    // The gamepads pointer (the ones received in the callbacks) gets updated
-    // automatically.
-    BP32.update();
+    // This call fetches all the controllers' data.
+    // Call this function in your main loop.
+    bool dataUpdated = BP32.update();
+    if (dataUpdated)
+        processControllers();
 
-    // It is safe to always do this before using the gamepad API.
-    // This guarantees that the gamepad is valid and connected.
-    for (int i = 0; i < BP32_MAX_GAMEPADS; i++) {
-        ControllerPtr myController = myControllers[i];
-
-        if (myController && myController->isConnected()) {
-            if (myController->isGamepad()) {
-                processGamepad(myController);
-            } else if (myController->isMouse()) {
-                processMouse(myController);
-            } else if (myController->isKeyboard()) {
-                processKeyboard(myController);
-            } else if (myController->isBalanceBoard()) {
-                processBalanceBoard(myController);
-            } else {
-                Console.printf("Data not available yet\n");
-                continue;
-            }
-            // See ArduinoController.h for all the available functions.
-        }
-    }
     // The main loop must have some kind of "yield to lower priority task" event.
-    // Otherwise the watchdog will get triggered.
+    // Otherwise, the watchdog will get triggered.
     // If your main loop doesn't have one, just add a simple `vTaskDelay(1)`.
     // Detailed info here:
     // https://stackoverflow.com/questions/66278271/task-watchdog-got-triggered-the-tasks-did-not-reset-the-watchdog-in-time
 
-    // vTaskDelay(1);
+    //     vTaskDelay(1);
     delay(150);
 }